Tests can be globbed. Test names now use :colon.
This implements many changes. To get a test relative to the skeleton dir, it must be prefixed with a colon (e.g. :mrs), otherwise it will be an absolute/relative filesystem path. The --list-profiles option of the R (regression-test) command now takes an optional parameter for restricting the listed profiles. Internally, paths are resolved to absolute paths, and the canonical name of the test is obtained more reliably.

Not yet done: Tests with actual colons in the name need a way to escape the colon.

Planned: The :test syntax can be extended to specify gold and skeletons manually, as in gold:skel, but something might need to be done to allow for filesystem paths, e.g. /path/to/gold:/path/to/skel.